+++ changelog.md
@@ -9,3 +9,4 @@
 * eachAny copes with iterator function returning sync
 * Remove unnecessary bind() on iterator function
 * Tests for chaining when passed promise
+* Tests for `this` binding being preserved along chain